add readthedocs files and clean code
