+++ docs/source/api_reference/core.rst
@@ -1,0 +1,8 @@
+======================================================
+``hipomap.core`` --- HipoMap core components
+======================================================
+
+.. automodule:: hipomap.core
+   :members:
+   :show-inheritance:
+
