+++ docs/source/api_reference/probmap.rst
@@ -1,0 +1,8 @@
+======================================================
+``hipomap.probmap`` --- HipoMap Prob components
+======================================================
+
+.. automodule:: hipomap.probmap
+   :members:
+   :show-inheritance:
+
