+++ docs/source/conf.py
@@ -8,16 +8,12 @@
 
 import os
 import sys
-import typing
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('../..'))
 sys.path.insert(1, os.path.abspath('../../hipomap'))
-sys.path.insert(2, os.path.abspath('../../hipomap/WSI_Preprocessing'))
-sys.path.insert(3, os.path.abspath('../../hipomap/WSI_Preprocessing/Preprocessing'))
-
 
 # -- Project information -----------------------------------------------------
 
@@ -35,8 +31,14 @@ release = '0.1'
 # ones.
 extensions = [
     'myst_parser',
-    'sphinx.ext.autodoc'
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.mathjax",
+    # "nbsphinx"
 ]
+
+napoleon_use_ivar = True
 
 source_suffix = {
     '.rst': 'restructuredtext',
@@ -57,7 +59,8 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
+master_doc = 'index'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
